refactor(table): extract Transaction and its builder into its own module
The Transaction struct, its two constructors and TransactionBuilder are the
entry point to the module and read better away from the manifest assembly they
feed.
