+++ rust/lance-table/src/transaction.rs
@@ -12,6 +12,7 @@
 //! For more details please refer to the
 //! [Transaction Specification](https://lance.org/format/table/transaction/#transaction-types).
 
+mod builder;
 mod conflicts;
 mod index_maintenance;
 mod row_version;
@@ -25,6 +26,7 @@ mod validate;
 #[cfg(test)]
 pub(crate) mod test_support;
 
+pub use builder::{Transaction, TransactionBuilder};
 pub use operation::{
     DataOverlayGroup, DataReplacementGroup, Operation, RewriteGroup, RewrittenIndex, UpdateMode,
     UpdatedFragmentOffsets,
@@ -52,7 +54,6 @@ use lance_core::datatypes::{
     LANCE_UNENFORCED_CLUSTERING_KEY_POSITION, LANCE_UNENFORCED_PRIMARY_KEY,
     LANCE_UNENFORCED_PRIMARY_KEY_POSITION,
 };
-use lance_core::deepsize::DeepSizeOf;
 use lance_core::{Error, Result};
 use lance_file::version::LanceFileVersion;
 use lance_io::object_store::ObjectStore;
@@ -62,87 +63,8 @@ use std::{
     collections::{HashMap, HashSet},
     sync::Arc,
 };
-use uuid::Uuid;
-
-/// A change to a dataset that can be retried
-///
-/// This contains enough information to be able to build the next manifest,
-/// given the current manifest.
-#[derive(Debug, Clone, DeepSizeOf, PartialEq)]
-pub struct Transaction {
-    /// The version of the table this transaction is based off of. If this is
-    /// the first transaction, this should be 0.
-    pub read_version: u64,
-    pub uuid: String,
-    pub operation: Operation,
-    pub tag: Option<String>,
-    pub transaction_properties: Option<Arc<HashMap<String, String>>>,
-}
-
-/// Add TransactionBuilder for flexibly setting option without using `mut`
-pub struct TransactionBuilder {
-    read_version: u64,
-    // uuid is optional for builder since it can autogenerate
-    uuid: Option<String>,
-    operation: Operation,
-    tag: Option<String>,
-    transaction_properties: Option<Arc<HashMap<String, String>>>,
-}
-
-impl TransactionBuilder {
-    pub fn new(read_version: u64, operation: Operation) -> Self {
-        Self {
-            read_version,
-            uuid: None,
-            operation,
-            tag: None,
-            transaction_properties: None,
-        }
-    }
-
-    pub fn uuid(mut self, uuid: String) -> Self {
-        self.uuid = Some(uuid);
-        self
-    }
-
-    pub fn tag(mut self, tag: Option<String>) -> Self {
-        self.tag = tag;
-        self
-    }
-
-    pub fn transaction_properties(
-        mut self,
-        transaction_properties: Option<Arc<HashMap<String, String>>>,
-    ) -> Self {
-        self.transaction_properties = transaction_properties;
-        self
-    }
-
-    pub fn build(self) -> Transaction {
-        let uuid = self
-            .uuid
-            .unwrap_or_else(|| Uuid::new_v4().hyphenated().to_string());
-        Transaction {
-            read_version: self.read_version,
-            uuid,
-            operation: self.operation,
-            tag: self.tag,
-            transaction_properties: self.transaction_properties,
-        }
-    }
-}
 
 impl Transaction {
-    pub fn new_from_version(read_version: u64, operation: Operation) -> Self {
-        TransactionBuilder::new(read_version, operation).build()
-    }
-
-    pub fn new(read_version: u64, operation: Operation, tag: Option<String>) -> Self {
-        TransactionBuilder::new(read_version, operation)
-            .tag(tag)
-            .build()
-    }
-
     fn fragments_with_ids<'a, T>(
         new_fragments: T,
         fragment_id: &'a mut u64,
